requirements.txt
numpy
scipy
matplotlib
ipython
MDAnalysis
pandas
sympy
nose
biopython
